strict-v1 p2: tighten transition continuity checks
